Refactor Google Drive connector to welcome google sheets
