+++ connectors/src/connectors/google_drive/temporal/activities.ts
@@ -1,153 +1,47 @@
-import fs from "fs/promises";
-import type { GaxiosResponse } from "googleapis-common";
-import { GaxiosError } from "googleapis-common";
-import StatsD from "hot-shots";
-import os from "os";
-import PQueue from "p-queue";
-
-import {
-  deleteFromDataSource,
-  MAX_DOCUMENT_TXT_LEN,
-  renderDocumentTitleAndContent,
-  upsertToDatasource,
-} from "@connectors/lib/data_sources";
-import { HTTPError } from "@connectors/lib/error";
-import { nango_client } from "@connectors/lib/nango_client";
-import mainLogger from "@connectors/logger/logger";
-import type { DataSourceConfig } from "@connectors/types/data_source_config";
-import type { GoogleDriveObjectType } from "@connectors/types/google_drive";
-import { FILE_ATTRIBUTES_TO_FETCH } from "@connectors/types/google_drive";
-const { NANGO_GOOGLE_DRIVE_CONNECTOR_ID = "google" } = process.env;
 import type { ModelId } from "@dust-tt/types";
-import { cacheWithRedis } from "@dust-tt/types";
 import { uuid4 } from "@temporalio/workflow";
 import type { drive_v3 } from "googleapis";
-import { google } from "googleapis";
-import { OAuth2Client } from "googleapis-common";
-import type { CreationAttributes } from "sequelize";
+import type { GaxiosResponse } from "googleapis-common";
+import type { OAuth2Client } from "googleapis-common";
+import { GaxiosError } from "googleapis-common";
+import StatsD from "hot-shots";
+import PQueue from "p-queue";
 import { literal, Op } from "sequelize";
 
 import { registerWebhook } from "@connectors/connectors/google_drive/lib";
+import { getGoogleDriveObject } from "@connectors/connectors/google_drive/lib/google_drive_api";
+import { getFileParentsMemoized } from "@connectors/connectors/google_drive/lib/hierarchy";
+import { syncOneFile } from "@connectors/connectors/google_drive/temporal/file";
+import { getMimesTypeToSync } from "@connectors/connectors/google_drive/temporal/mime_types";
+import {
+  driveObjectToDustType,
+  getAuthObject,
+  getDocumentId,
+  getDriveClient,
+} from "@connectors/connectors/google_drive/temporal/utils";
 import { dataSourceConfigFromConnector } from "@connectors/lib/api/data_source_config";
-import { dpdf2text } from "@connectors/lib/dpdf2text";
+import { deleteFromDataSource } from "@connectors/lib/data_sources";
+import { HTTPError } from "@connectors/lib/error";
 import { ExternalOauthTokenError } from "@connectors/lib/error";
 import {
-  GoogleDriveConfig,
   GoogleDriveFiles,
   GoogleDriveFolders,
   GoogleDriveSyncToken,
   GoogleDriveWebhook,
 } from "@connectors/lib/models/google_drive";
-import { getConnectionFromNango } from "@connectors/lib/nango_helpers";
 import { syncFailed } from "@connectors/lib/sync_status";
+import mainLogger from "@connectors/logger/logger";
 import logger from "@connectors/logger/logger";
 import { sequelizeConnection } from "@connectors/resources/storage";
 import { ConnectorModel } from "@connectors/resources/storage/models/connector_model";
+import type { DataSourceConfig } from "@connectors/types/data_source_config";
+import type { GoogleDriveObjectType } from "@connectors/types/google_drive";
+import { FILE_ATTRIBUTES_TO_FETCH } from "@connectors/types/google_drive";
 
 const FILES_SYNC_CONCURRENCY = 10;
 const FILES_GC_CONCURRENCY = 5;
 
-export const MIME_TYPES_TO_EXPORT: { [key: string]: string } = {
-  "application/vnd.google-apps.document": "text/plain",
-  "application/vnd.google-apps.presentation": "text/plain",
-};
-
-async function getMimeTypesToDownload(connectorId: ModelId) {
-  const mimeTypes = ["text/plain"];
-  const config = await GoogleDriveConfig.findOne({
-    where: {
-      connectorId: connectorId,
-    },
-  });
-  if (config?.pdfEnabled) {
-    mimeTypes.push("application/pdf");
-  }
-
-  return mimeTypes;
-}
-
-async function getMimesTypeToSync(connectorId: ModelId) {
-  const mimeTypes = await getMimeTypesToDownload(connectorId);
-  mimeTypes.push(...Object.keys(MIME_TYPES_TO_EXPORT));
-  mimeTypes.push("application/vnd.google-apps.folder");
-
-  return mimeTypes;
-}
-
 export const statsDClient = new StatsD();
-
-type NangoGetConnectionRes = {
-  connection_id: string;
-  credentials: {
-    type: string;
-    access_token: string;
-    refresh_token: string;
-    expires_at: string;
-    expires_in: number;
-    raw: {
-      scope: string;
-      token_type: string;
-    };
-  };
-};
-
-export async function getGoogleCredentials(
-  nangoConnectionId: string
-): Promise<NangoGetConnectionRes> {
-  if (!NANGO_GOOGLE_DRIVE_CONNECTOR_ID) {
-    throw new Error("NANGO_GOOGLE_DRIVE_CONNECTOR_ID is not defined");
-  }
-  return nango_client().getConnection(
-    NANGO_GOOGLE_DRIVE_CONNECTOR_ID,
-    nangoConnectionId,
-    false
-  );
-}
-
-export async function getAuthObject(
-  nangoConnectionId: string
-): Promise<OAuth2Client> {
-  if (!NANGO_GOOGLE_DRIVE_CONNECTOR_ID) {
-    throw new Error("NANGO_GOOGLE_DRIVE_CONNECTOR_ID is not defined");
-  }
-  const res: NangoGetConnectionRes = await getConnectionFromNango({
-    connectionId: nangoConnectionId,
-    integrationId: NANGO_GOOGLE_DRIVE_CONNECTOR_ID,
-    refreshToken: false,
-    useCache: true,
-  });
-
-  const oauth2Client = new google.auth.OAuth2();
-  oauth2Client.setCredentials({
-    access_token: res.credentials.access_token,
-    scope: res.credentials.raw.scope,
-    token_type: res.credentials.raw.token_type,
-    expiry_date: new Date(res.credentials.expires_at).getTime(),
-  });
-
-  return oauth2Client;
-}
-
-export async function getDriveClient(
-  auth_credentials: OAuth2Client
-): Promise<drive_v3.Drive>;
-export async function getDriveClient(
-  auth_credentials: string
-): Promise<drive_v3.Drive>;
-export async function getDriveClient(
-  auth_credentials: string | OAuth2Client
-): Promise<drive_v3.Drive> {
-  if (auth_credentials instanceof OAuth2Client) {
-    return google.drive({ version: "v3", auth: auth_credentials });
-  } else if (typeof auth_credentials === "string") {
-    const auth = await getAuthObject(auth_credentials);
-    const drive = google.drive({ version: "v3", auth: auth });
-
-    return drive;
-  }
-
-  throw new Error("Invalid auth_credentials type");
-}
 
 export async function getDrivesIds(connectorId: ModelId): Promise<
   {
@@ -310,364 +204,6 @@ export async function syncFiles(
     subfolders: subfolders,
   };
 }
-
-async function syncOneFile(
-  connectorId: ModelId,
-  oauth2client: OAuth2Client,
-  dataSourceConfig: DataSourceConfig,
-  file: GoogleDriveObjectType,
-  startSyncTs: number,
-  isBatchSync = false
-): Promise<boolean> {
-  const mimeTypesToDownload = await getMimeTypesToDownload(connectorId);
-  const documentId = getDocumentId(file.id);
-  let documentContent: string | undefined = undefined;
-
-  const fileInDb = await GoogleDriveFiles.findOne({
-    where: {
-      connectorId: connectorId,
-      driveFileId: file.id,
-    },
-  });
-
-  if (fileInDb?.skipReason) {
-    logger.info(
-      {
-        documentId,
-        dataSourceConfig,
-        fileId: file.id,
-        title: file.name,
-      },
-      `Google Drive document skipped with skip reason ${fileInDb.skipReason}`
-    );
-    return false;
-  }
-  if (!file.capabilities.canDownload) {
-    logger.info(
-      {
-        documentId,
-        connectorId,
-        fileId: file.id,
-        title: file.name,
-      },
-      `Google Drive document skipped because it cannot be downloaded`
-    );
-    return false;
-  }
-
-  if (MIME_TYPES_TO_EXPORT[file.mimeType]) {
-    const drive = await getDriveClient(oauth2client);
-    try {
-      const res = await drive.files.export({
-        fileId: file.id,
-        mimeType: MIME_TYPES_TO_EXPORT[file.mimeType],
-      });
-      if (res.status !== 200) {
-        logger.error(
-          {
-            documentId,
-            dataSourceConfig,
-            fileId: file.id,
-            title: file.name,
-          },
-          "Error exporting Google document"
-        );
-        throw new Error(
-          `Error exporting Google document. status_code: ${res.status}. status_text: ${res.statusText}`
-        );
-      }
-      if (typeof res.data === "string") {
-        documentContent = res.data;
-      } else if (
-        typeof res.data === "object" ||
-        typeof res.data === "number" ||
-        typeof res.data === "boolean" ||
-        typeof res.data === "bigint"
-      ) {
-        // In case the contents returned by the file export matches a JS type,
-        // we need to convert it
-        //  e.g. a google presentation with just the number
-        // 1 in it, the export will return the number 1 instead of a string
-        documentContent = res.data?.toString();
-      } else {
-        logger.error(
-          {
-            connectorId: connectorId,
-            documentId,
-            fileMimeType: file.mimeType,
-            fileId: file.id,
-            title: file.name,
-            resDataTypeOf: typeof res.data,
-            type: "export",
-          },
-          "Unexpected GDrive export response type"
-        );
-      }
-    } catch (e) {
-      logger.error(
-        {
-          documentId,
-          dataSourceConfig,
-          fileId: file.id,
-          title: file.name,
-        },
-        "Error exporting Google document"
-      );
-      throw e;
-    }
-  } else if (mimeTypesToDownload.includes(file.mimeType)) {
-    const drive = await getDriveClient(oauth2client);
-
-    let res;
-    try {
-      res = await drive.files.get(
-        {
-          fileId: file.id,
-          alt: "media",
-        },
-        {
-          responseType: "arraybuffer",
-        }
-      );
-    } catch (e) {
-      const maybeErrorWithCode = e as { code: string };
-      if (maybeErrorWithCode.code === "ERR_OUT_OF_RANGE") {
-        // This error happens when the file is too big to be downloaded.
-        // We skip this file.
-        logger.info(
-          {
-            file_id: file.id,
-            mimeType: file.mimeType,
-            title: file.name,
-          },
-          `File too big to be downloaded. Skipping`
-        );
-        return false;
-      }
-      throw e;
-    }
-
-    if (res.status !== 200) {
-      throw new Error(
-        `Error downloading Google document. status_code: ${res.status}. status_text: ${res.statusText}`
-      );
-    }
-
-    if (file.mimeType === "text/plain") {
-      if (res.data instanceof ArrayBuffer) {
-        // If data is > 4 times the limit, we skip the file since even if
-        // converted to utf-8 it will overcome the limit enforced below. This
-        // avoids operations on very long text files, that can cause
-        // Buffer.toString to crash if the file is > 500MB
-        if (res.data.byteLength > 4 * MAX_DOCUMENT_TXT_LEN) {
-          logger.info(
-            {
-              file_id: file.id,
-              mimeType: file.mimeType,
-              title: file.name,
-            },
-            `File too big to be chunked. Skipping`
-          );
-          return false;
-        }
-        documentContent = Buffer.from(res.data).toString("utf-8");
-      } else {
-        logger.error(
-          {
-            connectorId: connectorId,
-            documentId,
-            fileMimeType: file.mimeType,
-            fileId: file.id,
-            title: file.name,
-            resDataTypeOf: typeof res.data,
-            type: "download",
-          },
-          "Unexpected GDrive export response type"
-        );
-      }
-    } else if (file.mimeType === "application/pdf") {
-      const pdf_path = os.tmpdir() + "/" + uuid4() + ".pdf";
-      try {
-        if (res.data instanceof ArrayBuffer) {
-          await fs.writeFile(pdf_path, Buffer.from(res.data), "binary");
-        }
-
-        const pdfTextData = await dpdf2text(pdf_path);
-
-        documentContent = pdfTextData;
-        logger.info(
-          {
-            file_id: file.id,
-            mimeType: file.mimeType,
-            title: file.name,
-          },
-          `Successfully converted PDF to text`
-        );
-      } catch (err) {
-        logger.warn(
-          {
-            error: err,
-            file_id: file.id,
-            mimeType: file.mimeType,
-            filename: file.name,
-          },
-          `Error while converting PDF to text`
-        );
-        // we don't know what to do with PDF files that fails to be converted to text.
-        // So we log the error and skip the file.
-        return false;
-      } finally {
-        await fs.unlink(pdf_path);
-      }
-    }
-  } else {
-    // We do not support this file type
-    return false;
-  }
-
-  documentContent = documentContent?.trim();
-
-  const content = await renderDocumentTitleAndContent({
-    dataSourceConfig,
-    title: file.name,
-    updatedAt: file.updatedAtMs ? new Date(file.updatedAtMs) : undefined,
-    createdAt: file.createdAtMs ? new Date(file.createdAtMs) : undefined,
-    lastEditor: file.lastEditor ? file.lastEditor.displayName : undefined,
-    content: documentContent
-      ? { prefix: null, content: documentContent, sections: [] }
-      : null,
-  });
-
-  if (documentContent === undefined) {
-    logger.error(
-      {
-        connectorId: connectorId,
-        documentId,
-        fileMimeType: file.mimeType,
-        fileId: file.id,
-        title: file.name,
-      },
-      "documentContent is undefined"
-    );
-    throw new Error("documentContent is undefined");
-  }
-  const tags = [`title:${file.name}`];
-  if (file.updatedAtMs) {
-    tags.push(`updatedAt:${file.updatedAtMs}`);
-  }
-  if (file.createdAtMs) {
-    tags.push(`createdAt:${file.createdAtMs}`);
-  }
-  if (file.lastEditor) {
-    tags.push(`lastEditor:${file.lastEditor.displayName}`);
-  }
-  tags.push(`mimeType:${file.mimeType}`);
-
-  let upsertTimestampMs: number | undefined = undefined;
-
-  if (
-    documentContent.length > 0 &&
-    documentContent.length <= MAX_DOCUMENT_TXT_LEN
-  ) {
-    const parents = (
-      await getFileParentsMemoized(connectorId, oauth2client, file, startSyncTs)
-    ).map((f) => f.id);
-    parents.push(file.id);
-    parents.reverse();
-
-    await upsertToDatasource({
-      dataSourceConfig,
-      documentId,
-      documentContent: content,
-      documentUrl: file.webViewLink,
-      timestampMs: file.updatedAtMs,
-      tags,
-      parents: parents,
-      upsertContext: {
-        sync_type: isBatchSync ? "batch" : "incremental",
-      },
-    });
-
-    upsertTimestampMs = file.updatedAtMs;
-  } else {
-    logger.info(
-      {
-        documentId,
-        dataSourceConfig,
-        documentLen: documentContent.length,
-        title: file.name,
-      },
-      `Document is empty or too big to be upserted. Skipping`
-    );
-  }
-
-  const params: CreationAttributes<GoogleDriveFiles> = {
-    connectorId: connectorId,
-    dustFileId: documentId,
-    driveFileId: file.id,
-    name: file.name,
-    mimeType: file.mimeType,
-    parentId: file.parent,
-    lastSeenTs: new Date(),
-  };
-
-  if (upsertTimestampMs) {
-    params.lastUpsertedTs = new Date(upsertTimestampMs);
-  }
-
-  await GoogleDriveFiles.upsert(params);
-
-  return !!upsertTimestampMs;
-}
-
-// Please consider using the memoized version getFileParentsMemoized instead of this one.
-async function getFileParents(
-  connectorId: ModelId,
-  authCredentials: OAuth2Client,
-  driveFile: GoogleDriveObjectType,
-  /* eslint-disable @typescript-eslint/no-unused-vars */
-  startSyncTs: number
-): Promise<GoogleDriveObjectType[]> {
-  const logger = mainLogger.child({
-    provider: "google_drive",
-    connectorId: connectorId,
-  });
-  const parents: GoogleDriveObjectType[] = [];
-  let currentObject = driveFile;
-  while (currentObject.parent) {
-    const parent = await getGoogleDriveObject(
-      authCredentials,
-      currentObject.parent
-    );
-    if (!parent) {
-      // If we got a 404 error we stop the iteration as the parent disappeared.
-      logger.info("Parent not found in `getFileParents`", {
-        parentId: currentObject.parent,
-        fileId: driveFile.id,
-      });
-      break;
-    }
-    parents.push(parent);
-    currentObject = parent;
-  }
-
-  return parents.reverse();
-}
-
-export const getFileParentsMemoized = cacheWithRedis(
-  getFileParents,
-  (
-    connectorId: ModelId,
-    authCredentials: OAuth2Client,
-    driveFile: GoogleDriveObjectType,
-    startSyncTs: number
-  ) => {
-    const cacheKey = `${connectorId}-${startSyncTs}-${driveFile.id}`;
-
-    return cacheKey;
-  },
-  60 * 10 * 1000
-);
 
 export async function objectIsInFolders(
   connectorId: ModelId,
@@ -1162,96 +698,6 @@ async function deleteFile(googleDriveFile: GoogleDriveFiles) {
     await deleteFromDataSource(dataSourceConfig, googleDriveFile.dustFileId);
   }
   await googleDriveFile.destroy();
-}
-
-export async function getGoogleDriveObject(
-  authCredentials: OAuth2Client,
-  driveObjectId: string
-): Promise<GoogleDriveObjectType | null> {
-  const drive = await getDriveClient(authCredentials);
-
-  try {
-    const res = await drive.files.get({
-      fileId: driveObjectId,
-      supportsAllDrives: true,
-      fields: FILE_ATTRIBUTES_TO_FETCH.join(","),
-    });
-    if (res.status !== 200) {
-      throw new Error(
-        `Error getting files. status_code: ${res.status}. status_text: ${res.statusText}`
-      );
-    }
-    const file = res.data;
-
-    return await driveObjectToDustType(file, authCredentials);
-  } catch (e) {
-    if ((e as GaxiosError).response?.status === 404) {
-      return null;
-    }
-    throw e;
-  }
-}
-
-export async function driveObjectToDustType(
-  file: drive_v3.Schema$File,
-  authCredentials: OAuth2Client
-): Promise<GoogleDriveObjectType> {
-  if (
-    !file.name ||
-    !file.mimeType ||
-    !file.createdTime ||
-    !file.capabilities ||
-    file.capabilities.canDownload === undefined
-  ) {
-    throw new Error("Invalid file. File is: " + JSON.stringify(file));
-  }
-  const drive = await getDriveClient(authCredentials);
-  if (file.driveId == file.id) {
-    // We are dealing with a Google Drive object. We need a query to the Drive API to get the actual Drive name.
-    const driveRes = await drive.drives.get({
-      driveId: file.id as string,
-    });
-    if (driveRes.status !== 200) {
-      throw new Error(
-        `Error getting drive. status_code: ${driveRes.status}. status_text: ${driveRes.statusText}`
-      );
-    }
-    return {
-      id: driveRes.data.id as string,
-      name: driveRes.data.name as string,
-      parent: null,
-      mimeType: "application/vnd.google-apps.folder",
-      webViewLink: file.webViewLink ? file.webViewLink : undefined,
-      createdAtMs: new Date(file.createdTime).getTime(),
-      trashed: false,
-      capabilities: {
-        canDownload: false,
-      },
-    };
-  } else {
-    return {
-      id: file.id as string,
-      name: file.name,
-      parent: file.parents && file.parents[0] ? file.parents[0] : null,
-      mimeType: file.mimeType,
-      webViewLink: file.webViewLink ? file.webViewLink : undefined,
-      createdAtMs: new Date(file.createdTime).getTime(),
-      trashed: file.trashed ? file.trashed : false,
-      updatedAtMs: file.modifiedTime
-        ? new Date(file.modifiedTime).getTime()
-        : undefined,
-      lastEditor: file.lastModifyingUser
-        ? { displayName: file.lastModifyingUser.displayName as string }
-        : undefined,
-      capabilities: {
-        canDownload: file.capabilities.canDownload,
-      },
-    };
-  }
-}
-
-export function getDocumentId(driveFileId: string): string {
-  return `gdrive-${driveFileId}`;
 }
 
 export async function markFolderAsVisited(
